avatarapi/requirements.txt
PyPDF2
aiofiles
aiofiles==23.2.1
aiohttp
edge-tts
faiss-cpu
fastapi
fastapi==0.109.2
gTTS==2.5.1
googletrans==4.0.0-rc.1
langchain
langchain-community
motor
psycopg2-binary
pypdf
python-dotenv
python-jose[cryptography]
python-multipart
python-multipart==0.0.9
redis
requests
resend
sentence-transformers
sqlalchemy
uvicorn==0.27.1
uvicorn[standard]